added privacy page, moved pagebanner
