+++ src/pages/Careers.tsx
@@ -4,7 +4,7 @@ import ReactGA from 'react-ga4'
 import ReactPixel from 'react-facebook-pixel'
 
 // Component Imports
-import PageBanner from '../components/pagesections/PageBanner'
+import PageBanner from '../components/pagesections/banners/PageBanner'
 import BenefitsSection from '../components/pagesections/BenefitsSection'
 import ContactSection from '../components/pagesections/ContactSection'
 import CultureSection from '../components/pagesections/CultureSection'
