+++ src/pages/Community.tsx
@@ -4,7 +4,7 @@ import ReactGA from 'react-ga4'
 import ReactPixel from 'react-facebook-pixel'
 
 // Component Imports
-import PageBanner from '../components/pagesections/PageBanner'
+import PageBanner from '../components/pagesections/banners/PageBanner'
 import CommunitySection from '../components/pagesections/CommunitySection'
 import CommunityEventsSection from '../components/pagesections/CommunityEventsSection'
 import CommunityBiggestHeartSection from '../components/pagesections/CommunityBiggestHeartSection'
