+++ src/pages/Privacy.tsx
@@ -1,0 +1,9 @@
+import React, {useEffect} from 'react'
+
+const Privacy: React.FC = () => {
+
+    return (
+        <>
+        </>
+    )
+}
